app/requirements.txt
baidu-aip
python-dotenv
fastapi
numpy
opencv-python
uvicorn
